V0.2.0
Addition of the possibility to consider more general objective functions for PRIM. In particular, functions can now be used that take `x` as an argument. This can be useful if for instance the objective function involves a regression between the response and the predictors. See the help of `peeling` for more details.

Small bug fixes:
- Issue in the output when the number of steps reached the maximum number possible.
- cv.trajectory. Name of the argument `beta.stop.grid` changed to `grid` to avoid confusion in the calls to `peeling`.
- More general y label in the peeling trajectory plot.